requirements.txt
pillow==11.3.0
screeninfo==0.8.1
requests==2.32.5
